requirements_rag.txt
numpy>=1.26.0
faiss-cpu>=1.9.0
sentence-transformers>=2.2.2
datasets>=2.18.0
torch>=2.0.1
transformers>=4.6.0
huggingface-hub>=0.20.0
groq>=0.11.0
flask>=2.3.2
python-dotenv>=1.0.0
gunicorn>=20.1.0
flask-cors>=3.0.10
google-generativeai>=0.3.3